requirements.txt
arabic-reshaper==2.1.3
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.2
django-crispy-forms==1.10.0
django-filter==2.4.0
django-import-export==2.5.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
et-xmlfile==1.1.0
future==0.18.2
html5lib==1.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.1.0
pycodestyle==2.6.0
pylint==2.6.0
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2020.1
PyYAML==5.4.1
reportlab==3.6.1
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
urllib3==1.26.2
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
